Find average grade by year and department service

+++ src/main/java/org/unibl/etf/ps/cleanbl/dto/StatisticsDTO.java
@@ -11,5 +11,5 @@ import java.util.List;
 @NoArgsConstructor
 public class StatisticsDTO {
     private String name;
-    private List<Integer> data;
+    private List<Double> data;
 }
